refactor instance heart beat storage
